Create chi2ind.py
The Chi-square independence test is use to determine if two categorical variables are related or not. This is a useful procedure to go through when preprocessing categorical data for Data Science tasks. We found the below link a good resource to develop an understanding of Chi-square independence test

https://www.spss-tutorials.com/chi-square-independence-test/

The Python code provided here takes a data frame as an input expecting at least two columns of categorical variables in it and produces two data frames, one of P-values (to indicate to the significance) and the other of Cramers V values (to indicate to the strength of a relationship).